% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline-package.R
\docType{package}
\name{timeline-package}
\alias{timeline-package}
\title{timeline}
\description{
Package to create timelines.
}
\author{
Jason Bryer \email{jason@bryer.org}
}
\keyword{ggplot2}
\keyword{timeline}

